feat: stub for instant resharding state root creation (#12083)
Create the code flow intended for resharding state root creation.
The ultimate goal will be the temporary memtries created, which will be
used until new memtries are created from flat storage.

This will be a multi-step work. To complete it, all mentioned TODOs must
be resolved.

### Overview

The core is a `MemTrieUpdate::retain_multi_range` function and a simple
`test_retain_multi_range` test. It follows the logic in #12074. I plan
to enhance logic sequentially until it will be fit for the actual
resharding.
Drawback is that code is not used on chain, but setting up full
resharding test requires separate work. I suggest to start from the the
incomplete flow; hopefully I will need ~5 PRs to set up the full test.

It follows the "top-down" logic as in insert/delete functions: first we
pull root node to in-memory update list, then we descend into children.
In the end, we call `to_trie_changes` which calls
`post_order_traverse_updated_nodes` to create full set of changes to be
applied to trie.

### Alternative

For the multi-range deletion, better logic is "bottom-up": we first
understand the result of children subtrees, and then mark root node as
updated if needed.
It's not clear whether everything must be "bottom-up". It probably has
its own challenges. But here it is strictly better because it
automatically gives the post traversal order of updated nodes.
We could compute hashes on the fly but it's less LoC to postpone it to
reuse existing logic, so I just call
`compute_hashes_and_serialized_nodes` in the end.