fix(network): bound size of tier3_requests queue (#12124)
This simple limit is added for security purposes; it prevents the node
from going OOM due to unbounded addition to the queue. In the future we
will implement separate resource pools to prioritize incoming requests
from validator nodes.

--- 

Currently a node processes at most one tier3 request per second:


https://github.com/near/nearcore/blob/d6d6e4782f7d57a7b7acc2894e1668355012ff74/chain/network/src/peer_manager/peer_manager_actor.rs#L92

and the default timeout on the node which sent the request is 60
seconds:


https://github.com/near/nearcore/blob/d6d6e4782f7d57a7b7acc2894e1668355012ff74/core/chain-configs/src/client_config.rs#L282

Hence a queue limit of 60 means that any dropped requests would not have
been served in time anyway.