[memtrie] Fix refcount for HybridArena (#12118)
On converting existing arena memory to shared memory, we lose the
ability to refcount the shared memory part.

However, while adding new nodes and new entries in the owned memory
part, we could potentially be referring to the shared memory.

In such cases, while adding or removing nodes, we would like to ignore
all increments and decrements to refcounts in the shared part of memory.

Added a unit test for expected scenario.