fix(state-snapshot): correctly return included shard uids (#12143)
The StateSnapshot class provides a function called `get_shard_uids`,
which is intended to return the UIDs of the shards included in the
snapshot.

However, this function was incorrectly returning the entire set of flat
storage keys from the flat storage manager. As a result, if a flat
storage exists but the snapshot process fails to update the flat storage
head, the snapshot would incorrectly report that the shard was
successfully included.