[resharding] Introduce ReshardingManager (#12149)
The resharding manager should hopefully be a one stop shop for all
things resharding.

For now I'm keeping it as a member of Chain instead of Client but we can
change that later.

The resharding manager in the future should be able to do the following
work
- Figure out when to schedule a resharding event based on next epoch
info, last block of epoch, shard layout etc.
- Manage all the sync preprocessing work like
  - Split parent trie into two parts
  - Update ShardTries with info about the two new child tries
- Send request to any async job (probably we'll need a ReshardingActor)
- Manage all the sync postprocessing work like
  - Replace child tries with the newly created tries
  

There are some complications related to accessing store components that
I plan to simplify in the future