fix(tests): fix state sync test errors (#12150)
These tests create node home dirs in tempfiles, but these are dropped
too soon, and the directories are removed while the nodes are still
running. This is visible in the logs as snapshot creation failure
messages. After https://github.com/near/nearcore/pull/12147, these no
longer cause test failures, but the underlying failures to create
snapshots are still there.

Fix it by keeping them around in an Arc in the enclosing scopes

Note that every now and then these tests still fail with many messages
showing `Received an invalid block during state sync` followed by a test
timeout, but it seems to be for an unrelated reason.