test: enable simple resharding v3 test (#12191)
Do minimal changes to the code allowing to check that number of shards
in the new epoch increased.

This code can be reused to test each separate component for resharding
v3:
* non-contiguous shard ids
* state sync
* memtries resharding
* ...

To make the test pass, nodes must track all shards for now, because
state sync is not implemented yet. So every node must think that it has
enough state to skip state sync.

Note that it doesn't mean at all that resharding works already. State is
also not properly constructed yet, so tx processing will either be
incorrect or crash the node.