test: produce chunks in resharding v3 test (#12196)
Improve the test by checking that all chunks are included, and the chain
passes through one epoch after resharding.

For that, I need to properly skip chunk validation if resharding
happened in the middle, as this is not a part of early MVP. One more
useful and necessary change: on iterating over block range of state
transition, I need to move from shard_id to prev_shard_id, so that we
query proper chunk in a block.

Checked offline that without validation skip, chunks for shards 1-2 are
not included.