feat(resharding) - Make shard ids non-contiguous - part 3 (#12214)
The next part of refactoring ShardId to newtype. 

Organized by commits. The only non-trivial one should be the `state
snapshot refactoring`. The interface in there is ugly because the state
snapshot gets epoch manager captured inside of a lambda instead of
directly. I didn't have the patience to properly refactor it, so I just
added the shard indices to the lambda.