test: Add nayduck test for kicking out offline validators (#12215)
We add a simple integration test for checking that offline nodes are
kicked out properly. We kill a block+chunk producer and a
chunk-validator only node. Assert that these two nodes are kicked out
due to no block and endorsement production and not included in the next
validators.

This is part of the testing tasks for #11900.