test: warmup before tx execution (#12226)
https://nayduck.nearone.org/#/test/117339 was flaky on some run.
The failure happens because of ShardStuck error for the single shard we
have.
Looks like it is specific to the height 3, where there is a high
probability that no chunks can be produced which triggers ShardStuck.
But if we wait for height 5, everything is fine.

Another finding: sometimes `load_test_contract` seems to serve as an
unintended delay to ensure warmup. In some of the cases we do
`time.sleep` or `wait_for_blocks` as well.