[Epoch Sync] Add a Nayduck test for Epoch Sync. (#12237)
Most of the complexity here is changes to the StoreValidator:
* Do not check that previous block exists, if it's older than the epoch
sync boundary.
* Do not check that if BlockInfo exists the header also exists, because
this is also not true for some blocks before the epoch sync boundary.
* (This one was pre-existing before epoch sync) Be lenient when checking
chunk body's existence, because we don't seem to download these for the
extra blocks needed by state sync. I couldn't figure out a good way to
check this rigorously, so I included a heuristic to skip checking if the
block in question is <= tail height. I don't think it's correct, but it
works for now.
* The reason why this does not trigger state sync nayduck test failures
is because the utils.poll_blocks function doesn't trigger store
validations, but utils.wait_for_blocks does :roll_eyes: