feat(resharding): put resharding transition to state witness (#12262)
Add resharding transition to the list of `implicit_transitions` in the
state witness.
The idea is that shard split is simply a transition from one shard id to
another, and only the `PartialState` is enough to execute that.
Additionally, we need to move to previous shard id in
`collect_state_transition_data` and `validate_source_receipt_proofs`.
This and #12253 allow to uncomment tests related to skipped chunks. Note
that state witness including resharding transition is still considered
always valid; this work is needed in order not to crash during its
construction with error "some shard id doesn't exist for this layout"