feat: Add new unittests for apply function when contract code is excluded from state witness (#12294)
We also do some refactoring to 
1) allow creating multiple accounts in the initial state (otherwise
apply for non-existent account fails silently). for example some tests
are on both `allice` and `bob` but only `allice` account is created and
the receipts on the `bob` account fail without throwing an error.
2) make `receipt_id` unique by adding account id to the hash, otherwise
receipt preparation pipeline fails in the case of different accounts
running exactly the same receipts.