feat(epoch config): refactor epoch config to support test configs (#12283)
This PR introduces the option to load epoch config for each protocol
version from external files.
The new order in which we will load the epoch config is:
If mainnet or testnet => go for hardcoded json files.
If there is a non null `epoch_config` path provided and the directory
exists => load the files from that folder.
Otherwise build the epoch config from the genesis.

We kept the last option to keep the compatibility with the existing
tests.